Feature 343 deleted metaforms

+++ src/main/kotlin/fi/metatavu/metaform/server/rest/MetaformsApi.kt
@@ -105,19 +105,28 @@ class MetaformsApi: fi.metatavu.metaform.api.spec.MetaformsApi, AbstractApi() {
     }
   }
 
-  override fun deleteMetaform(metaformId: UUID): Response {
+  override fun deleteMetaform(metaformId: UUID, immediate: Boolean?): Response {
     loggedUserId ?: return createForbidden(UNAUTHORIZED)
+
+    if (immediate == true && environment != "DEVELOPMENT") {
+      return createForbidden(createNotAllowedMessage(DELETE, METAFORM))
+    }
 
     if (!isMetaformAdmin(metaformId)) {
       return createForbidden(createNotAllowedMessage(DELETE, METAFORM))
     }
 
-    val metaform = metaformController.findMetaformById(metaformId)
+    val metaform = metaformController.findMetaformById(metaformId, includeDeleted = true)
       ?: return createNotFound(createNotFoundMessage(METAFORM, metaformId))
 
     metaformScriptController.deleteMetaformScriptsByMetaform(metaform)
 
-    metaformController.deleteMetaform(metaform)
+    if (immediate == true) {
+      metaformController.deleteMetaform(metaform)
+    } else {
+      metaformController.updateMetaformDeleted(metaform)
+    }
+
 
     return createNoContent()
   }
@@ -138,6 +147,13 @@ class MetaformsApi: fi.metatavu.metaform.api.spec.MetaformsApi, AbstractApi() {
         ?: return createNotFound(createNotFoundMessage(METAFORM, metaformId))
     } else {
       return createBadRequest("Invalid request")
+    }
+
+    if (metaform.deleted!!) {
+      if (metaformSlug != null) {
+        return createNotFound(createSlugNotFoundMessage(METAFORM, metaformSlug))
+      }
+      return createNotFound(createNotFoundMessage(METAFORM, metaformId!!))
     }
 
     val translatedMetaform = try {
